+++ AndroidManifest.xml
