+++ DeviceInfoModel.java
