+++ DeviceListAdapter.java
