--- PressureSensorApp/.gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,0 @@
-#Sat Apr 02 15:22:20 EDT 2022
-gradle.version=7.2
