+++ PressureSensorApp/local.properties
@@ -1,8 +1,10 @@
-## This file must *NOT* be checked into Version Control Systems,
+## This file is automatically generated by Android Studio.
+# Do not modify this file -- YOUR CHANGES WILL BE ERASED!
+#
+# This file should *NOT* be checked into Version Control Systems,
 # as it contains information specific to your local configuration.
 #
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Sat Apr 02 15:32:21 EDT 2022
-sdk.dir=C\:\\Users\\andre\\AppData\\Local\\Android\\Sdk
+sdk.dir=/Users/ayahamer/Library/Android/sdk
