+++ SelectDeviceActivity.java
