+++ bluetooth_setup.xml
