+++ select_device_activity.xml
